+++ src/main/groovy/org/jetbrains/intellij/tasks/PublishTask.groovy
@@ -17,6 +17,7 @@ class PublishTask extends ConventionTask {
     private Object host = "https://plugins.jetbrains.com"
     private Object username
     private Object password
+    private Object token
     private List<Object> channels = new ArrayList<Object>()
 
     PublishTask() {
@@ -50,6 +51,21 @@ class PublishTask extends ConventionTask {
     }
 
     @Input
+    @Optional
+    String getToken() {
+        Utils.stringInput(token)
+    }
+
+    void setToken(Object token) {
+        this.token = token
+    }
+
+    void token(Object token) {
+        this.token = token
+    }
+
+    @Input
+    @Optional
     String getUsername() {
         Utils.stringInput(username)
     }
@@ -63,6 +79,7 @@ class PublishTask extends ConventionTask {
     }
 
     @Input
+    @Optional
     String getPassword() {
         Utils.stringInput(password)
     }
@@ -95,6 +112,7 @@ class PublishTask extends ConventionTask {
     @SuppressWarnings("GroovyUnusedDeclaration")
     @TaskAction
     protected void publishPlugin() {
+        validateInput()
         def channels = getChannels()
         if (!channels || channels.length == 0) {
             channels = ['default']
@@ -108,12 +126,13 @@ class PublishTask extends ConventionTask {
             for (String channel : channels) {
                 IntelliJPlugin.LOG.info("Uploading plugin ${pluginId} from $distributionFile.absolutePath to $host, channel: $channel")
                 try {
-                    def repoClient = new PluginRepositoryInstance(host, getUsername(), getPassword())
+                    def repoClient = getToken() ? new PluginRepositoryInstance(host, getToken())
+                            : new PluginRepositoryInstance(host, getUsername(), getPassword())
                     repoClient.uploadPlugin(pluginId, distributionFile, channel && 'default' != channel ? channel : '')
                     IntelliJPlugin.LOG.info("Uploaded successfully")
                 }
                 catch (exception) {
-                    throw new TaskExecutionException(this, new RuntimeException("Failed to upload plugin", exception))
+                    throw new TaskExecutionException(this, new GradleException("Failed to upload plugin", exception))
                 }
             }
         } else if (creationResult instanceof PluginCreationFail) {
@@ -121,6 +140,21 @@ class PublishTask extends ConventionTask {
             throw new TaskExecutionException(this, new GradleException("Cannot upload plugin. $problems"))
         } else {
             throw new TaskExecutionException(this, new GradleException("Cannot upload plugin. $creationResult"))
+        }
+    }
+
+    private void validateInput() {
+        if (!getToken()) {
+            if (!getPassword() && !getUsername()) {
+                throw new TaskExecutionException(this, new GradleException('token or username/password properties must be specified for plugin publishing'))
+            } else {
+                IntelliJPlugin.LOG.warn("Password authentication for uploading is deprecated. Use token access instead: http://www.jetbrains.org/intellij/sdk/docs/plugin_repository/api/plugin_upload.html")
+                if (!getPassword()) {
+                    throw new TaskExecutionException(this, new GradleException('No value has been specified for property \'password\''))
+                } else if (!getUsername()) {
+                    throw new TaskExecutionException(this, new GradleException('No value has been specified for property \'username\''))
+                }
+            }
         }
     }
 }